HTBHF-2520 move common paths out of steps directory
