+++ src/web/routes/application/flow-control/middleware/handle-path-request/selectors.js
@@ -1,4 +1,4 @@
-const { CHECK_ANSWERS_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL } = require('../../../steps/common/constants')
+const { CHECK_ANSWERS_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL } = require('../../../paths')
 
 const getPathsInSequence = (steps) => [...steps.map(step => step.path), CHECK_ANSWERS_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL]
