+++ src/web/routes/application/flow-control/middleware/handle-path-request/selectors.test.js
@@ -1,5 +1,5 @@
 const test = require('tape')
-const { CHECK_ANSWERS_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL } = require('../../../steps/common/constants')
+const { CHECK_ANSWERS_URL, TERMS_AND_CONDITIONS_URL, CONFIRM_URL } = require('../../../paths')
 const { getPathsInSequence } = require('./selectors')
 
 const steps = [{ path: '/first', next: () => '/second' }, { path: '/second' }]
