+++ src/web/routes/application/paths.js
@@ -1,0 +1,3 @@
+module.exports.CHECK_ANSWERS_URL = '/check-answers'
+module.exports.CONFIRM_URL = '/confirm'
+module.exports.TERMS_AND_CONDITIONS_URL = '/terms-and-conditions'
