+++ src/web/routes/application/steps/check-answers/check-answers.js
@@ -1,5 +1,5 @@
 const { getCheckAnswers } = require('./get')
-const { CHECK_ANSWERS_URL } = require('../common/constants')
+const { CHECK_ANSWERS_URL } = require('../../paths')
 const { handleRequestForPath } = require('../../flow-control')
 
 const registerCheckAnswersRoutes = (journey, config, app) => {
