+++ src/web/routes/application/steps/common/constants.js
@@ -1,9 +1,5 @@
 module.exports.YES = 'yes'
 module.exports.NO = 'no'
-module.exports.CHECK_ANSWERS_URL = '/check-answers'
-module.exports.CONFIRM_URL = '/confirm'
-module.exports.TERMS_AND_CONDITIONS_URL = '/terms-and-conditions'
-module.exports.UNSUCCESSFUL_APPLICATION = '/unsuccessful-application'
 module.exports.ELIGIBLE = 'ELIGIBLE'
 module.exports.EMAIL = 'email'
 module.exports.TEXT = 'text'
