+++ src/web/routes/application/steps/terms-and-conditions/post.test.js
@@ -2,7 +2,8 @@ const test = require('tape')
 const sinon = require('sinon')
 const proxyquire = require('proxyquire')
 const { states } = require('../../flow-control')
-const { CHECK_ANSWERS_URL, CONFIRM_URL, ELIGIBLE } = require('../common/constants')
+const { CHECK_ANSWERS_URL, CONFIRM_URL } = require('../../paths')
+const { ELIGIBLE } = require('../common/constants')
 
 const post = sinon.stub()
